modify: move to routes and connect AppPresenter with
